etl/requirements.txt

pandas
requests
python-dotenv
